Update notebooks to fresh env
