+++ py/requirements.txt
@@ -1,12 +1,12 @@
-ipykernel==6.29.2
-numpy==1.26.3
-matplotlib==3.8.2
-seaborn==0.13.1
-pandas==2.1.4
-pyarrow==14.0.2
-scikit-learn==1.3.2
-xgboost==2.0.3
-lightgbm==4.2.0
-shap==0.44.0
-tensorflow==2.15.0
-glum==3.0.0
+ipykernel==6.29.5
+numpy==1.26.4
+matplotlib==3.9.1
+seaborn==0.13.2
+pandas==2.2.2
+pyarrow==17.0.0
+scikit-learn==1.5.1
+xgboost==2.1.1
+lightgbm==4.5.0
+shap==0.46.0
+tensorflow==2.17.0
+glum==3.0.2
